+++ packages/fiori-mcp-server/test/unit/server.test.ts
@@ -23,10 +23,14 @@ describe('FioriFunctionalityServer', () => {
         setRequestHandlerMock.mockReset();
     });
 
+    // version cannot be hard coded as it will update on each new patch update
     test('Constructor', () => {
         new FioriFunctionalityServer();
         // Check initialization
-        expect(Server).toHaveBeenCalledWith({ name: 'fiori-mcp', version: '0.0.0' }, { capabilities: { tools: {} } });
+        expect(Server).toHaveBeenCalledWith(
+            { name: 'fiori-mcp', version: expect.any(String) },
+            { capabilities: { tools: {} } }
+        );
         expect(setRequestHandlerMock).toHaveBeenCalledTimes(2);
     });
 
@@ -49,15 +53,17 @@ describe('FioriFunctionalityServer', () => {
                 applications: [
                     {
                         name: 'app1',
-                        path: 'appPath1',
-                        type: 'list-report',
-                        version: '4.0'
+                        appPath: 'appPath1',
+                        projectType: 'EDMXBackend',
+                        projectPath: 'appPath1',
+                        odataVersion: '4.0'
                     },
                     {
                         name: 'app2',
-                        path: 'appPath2',
-                        type: 'list-report',
-                        version: '4.0'
+                        appPath: 'appPath2',
+                        projectType: 'EDMXBackend',
+                        projectPath: 'appPath2',
+                        odataVersion: '4.0'
                     }
                 ]
             });
@@ -78,15 +84,17 @@ describe('FioriFunctionalityServer', () => {
                 applications: [
                     {
                         name: 'app1',
-                        path: 'appPath1',
-                        type: 'list-report',
-                        version: '4.0'
+                        appPath: 'appPath1',
+                        projectType: 'EDMXBackend',
+                        projectPath: 'appPath1',
+                        odataVersion: '4.0'
                     },
                     {
                         name: 'app2',
-                        path: 'appPath2',
-                        type: 'list-report',
-                        version: '4.0'
+                        appPath: 'appPath2',
+                        projectType: 'EDMXBackend',
+                        projectPath: 'appPath2',
+                        odataVersion: '4.0'
                     }
                 ]
             });
@@ -103,11 +111,11 @@ describe('FioriFunctionalityServer', () => {
                 applicationPath: 'app1',
                 functionalities: [
                     {
-                        id: 'add-page',
+                        functionalityId: 'add-page',
                         description: 'Add page...'
                     },
                     {
-                        id: 'delete-page',
+                        functionalityId: 'delete-page',
                         description: 'Delete page...'
                     }
                 ]
@@ -130,11 +138,11 @@ describe('FioriFunctionalityServer', () => {
                 functionalities: [
                     {
                         description: 'Add page...',
-                        id: 'add-page'
+                        functionalityId: 'add-page'
                     },
                     {
                         description: 'Delete page...',
-                        id: 'delete-page'
+                        functionalityId: 'delete-page'
                     }
                 ]
             });
@@ -148,7 +156,7 @@ describe('FioriFunctionalityServer', () => {
 
         test('get-functionality-details', async () => {
             const getFunctionalityDetailsSpy = jest.spyOn(tools, 'getFunctionalityDetails').mockResolvedValue({
-                id: 'add-page',
+                functionalityId: 'add-page',
                 description: 'Add page...',
                 name: 'add-page',
                 parameters: []
@@ -169,7 +177,7 @@ describe('FioriFunctionalityServer', () => {
             const structuredContent = result.structuredContent;
             expect(structuredContent).toEqual({
                 description: 'Add page...',
-                id: 'add-page',
+                functionalityId: 'add-page',
                 name: 'add-page',
                 parameters: []
             });