+++ packages/odata-service-inquirer/test/unit/prompts/prompts.test.ts
@@ -4,6 +4,7 @@ import { getQuestions } from '../../../src/prompts';
 import { DatasourceType } from '../../../src/types';
 import * as utils from '../../../src/utils';
 import { hostEnvironment } from '@sap-ux/fiori-generator-shared';
+import { isFeatureEnabled } from '@sap-ux/feature-toggle';
 
 /**
  * Workaround to for spyOn TypeError: Jest cannot redefine property
@@ -14,6 +15,10 @@ jest.mock('@sap-ux/btp-utils', () => {
         ...jest.requireActual('@sap-ux/btp-utils')
     };
 });
+
+jest.mock('@sap-ux/feature-toggle', () => ({
+    isFeatureEnabled: jest.fn()
+}));
 
 jest.mock('@sap-ux/store', () => ({
     __esModule: true, // Workaround to for spyOn TypeError: Jest cannot redefine property
@@ -38,6 +43,10 @@ describe('getQuestions', () => {
     beforeAll(async () => {
         // Wait for i18n to bootstrap so we can test localised strings
         await initI18nOdataServiceInquirer();
+    });
+
+    beforeEach(() => {
+        (isFeatureEnabled as jest.Mock).mockReturnValue(false);
     });
 
     afterEach(() => {