+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.28
+
+### Patch Changes
+
+-   Updated dependencies [b3e95fb]
+    -   @sap-ux/fe-fpm-writer@0.38.2
+
+## 0.1.27
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.38.1
+
 ## 0.1.26
 
 ### Patch Changes