+++ packages/fiori-mcp-server/CHANGELOG.md
@@ -1,1 +1,20 @@
 # @sap-ux/fiori-mcp-server
+
+## 0.0.2
+
+### Patch Changes
+
+-   b2effee: fix CAP preview after generation - do npm install on CAP project and make service path absolute
+
+## 0.0.1
+
+### Patch Changes
+
+-   38ecd39: Initial version of the fiori-mcp-server package
+
+## 0.0.1
+
+### Patch Changes
+
+-   @sap-ux/fiori-annotation-api@0.6.17
+-   @sap-ux/project-access@1.30.13