+++ packages/nodejs-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/nodejs-utils
 
+## 0.2.3
+
+### Patch Changes
+
+-   Updated dependencies [4cfebaf]
+    -   @sap-ux/btp-utils@1.1.1
+
 ## 0.2.2
 
 ### Patch Changes